some changes on  footer, contact pages
